Move `inlineDataParts` tests to new `GenerateContentResponseTests` file
